restructure project, group common files. Hopefully make it a little more navigable
