+++ src/components/MarkdownRenderer.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import ReactMarkdown from 'react-markdown';
 
-import { MarkdownRendererProps } from '../types.tsx';
+import { MarkdownRendererProps } from '../types';
 
 
 const MarkdownRenderer: React.FC<MarkdownRendererProps> = ({ content }) => <ReactMarkdown>{content}</ReactMarkdown>;
