+++ src/types/image.d.ts
