+++ src/types/index.ts
@@ -1,0 +1,3 @@
+export * from './components';
+export * from './quiz';
+export * from './pages';
