+++ src/types/pages.ts
@@ -1,0 +1,9 @@
+export interface SelectCategoryProps {
+  selectQuiz: (category: string, index: number) => void;
+  startRandomQuiz: () => void;
+}
+
+export interface SelectQuestionsTotalProps {
+  startQuiz: (e: number) => void;
+  totalQuestions: number; // Add the totalQuestions prop
+}
