+++ src/types/quiz.ts
@@ -1,28 +1,4 @@
 import { MouseEventHandler } from 'react';
-import { NavLinkProps } from 'react-router-dom';
-
-export interface ButtonProps {
-  text: string;
-  isTransparent: boolean;
-  size: string;
-  handleClick: MouseEventHandler;
-}
-
-export interface ButtonLinkProps extends NavLinkProps {
-  size?: string;
-}
-
-export interface FlashcardProps {
-  front: string;
-  back: string;
-  trackProgress?: boolean;
-  known?: boolean;
-  onToggleKnown?: () => void;
-}
-
-export interface MarkdownRendererProps {
-  content: string;
-}
 
 export interface Quiz {
   category: string;
@@ -85,14 +61,4 @@ export interface PointTotals {
   points: number;
   totalQuestions: number;
   resetQuiz: () => void;
-}
-
-export interface SelectCategoryProps {
-  selectQuiz: (category: string, index: number) => void;
-  startRandomQuiz: () => void;
-}
-
-export interface SelectQuestionsTotalProps {
-  startQuiz: (e: number) => void;
-  totalQuestions: number; // Add the totalQuestions prop
 }
